move NibblesIterator to path.rs
