+++ storage/src/node/path.rs
@@ -5,10 +5,13 @@
 use bitflags::bitflags;
 use serde::{Deserialize, Serialize};
 use smallvec::SmallVec;
+use std::iter::FusedIterator;
 use std::{
     fmt::{self, Debug},
     iter::once,
 };
+
+static NIBBLES: [u8; 16] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15];
 
 /// Path is part or all of a node's path in the trie.
 /// Each element is a nibble.
@@ -103,7 +106,9 @@ impl Path {
     /// Create an iterator that returns the bytes from the underlying nibbles
     /// If there is an odd nibble at the end, it is dropped
     pub fn bytes_iter(&self) -> BytesIterator<'_> {
-        BytesIterator { iter: self.iter() }
+        BytesIterator {
+            nibbles_iter: self.iter(),
+        }
     }
 
     /// Create a boxed set of bytes from the Path
@@ -112,16 +117,18 @@ impl Path {
     }
 }
 
+/// Returns the nibbles in `nibbles_iter` as compressed bytes.
+/// That is, each two nibbles are combined into a single byte.
 #[derive(Debug)]
 pub struct BytesIterator<'a> {
-    iter: std::slice::Iter<'a, u8>,
+    nibbles_iter: std::slice::Iter<'a, u8>,
 }
 
 impl Iterator for BytesIterator<'_> {
     type Item = u8;
     fn next(&mut self) -> Option<Self::Item> {
-        if let Some(&hi) = self.iter.next() {
-            if let Some(&lo) = self.iter.next() {
+        if let Some(&hi) = self.nibbles_iter.next() {
+            if let Some(&lo) = self.nibbles_iter.next() {
                 return Some(hi * 16 + lo);
             }
         }
@@ -131,16 +138,158 @@ impl Iterator for BytesIterator<'_> {
     // this helps make the collection into a box faster
     fn size_hint(&self) -> (usize, Option<usize>) {
         (
-            self.iter.size_hint().0 / 2,
-            self.iter.size_hint().1.map(|max| max / 2),
+            self.nibbles_iter.size_hint().0 / 2,
+            self.nibbles_iter.size_hint().1.map(|max| max / 2),
         )
     }
 }
 
+/// Iterates over the nibbles in `data`.
+/// That is, each byte in `data` is converted to two nibbles.
+#[derive(Clone, Debug)]
+pub struct NibblesIterator<'a> {
+    data: &'a [u8],
+    head: usize,
+    tail: usize,
+}
+
+impl<'a> FusedIterator for NibblesIterator<'a> {}
+
+impl<'a> Iterator for NibblesIterator<'a> {
+    type Item = u8;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.is_empty() {
+            return None;
+        }
+        let result = if self.head % 2 == 0 {
+            #[allow(clippy::indexing_slicing)]
+            NIBBLES[(self.data[self.head / 2] >> 4) as usize]
+        } else {
+            #[allow(clippy::indexing_slicing)]
+            NIBBLES[(self.data[self.head / 2] & 0xf) as usize]
+        };
+        self.head += 1;
+        Some(result)
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let remaining = self.tail - self.head;
+        (remaining, Some(remaining))
+    }
+
+    fn nth(&mut self, n: usize) -> Option<Self::Item> {
+        self.head += std::cmp::min(n, self.tail - self.head);
+        self.next()
+    }
+}
+
+impl<'a> NibblesIterator<'a> {
+    #[inline(always)]
+    const fn is_empty(&self) -> bool {
+        self.head == self.tail
+    }
+
+    /// Returns a new `NibblesIterator` over the given `data`.
+    /// Each byte in `data` is converted to two nibbles.
+    pub const fn new(data: &'a [u8]) -> Self {
+        NibblesIterator {
+            data,
+            head: 0,
+            tail: 2 * data.len(),
+        }
+    }
+}
+
+impl<'a> DoubleEndedIterator for NibblesIterator<'a> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        if self.is_empty() {
+            return None;
+        }
+
+        let result = if self.tail % 2 == 0 {
+            #[allow(clippy::indexing_slicing)]
+            NIBBLES[(self.data[self.tail / 2 - 1] & 0xf) as usize]
+        } else {
+            #[allow(clippy::indexing_slicing)]
+            NIBBLES[(self.data[self.tail / 2] >> 4) as usize]
+        };
+        self.tail -= 1;
+
+        Some(result)
+    }
+
+    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
+        self.tail -= std::cmp::min(n, self.tail - self.head);
+        self.next_back()
+    }
+}
+
 #[cfg(test)]
+#[allow(clippy::indexing_slicing)]
 mod test {
     use super::*;
+    use std::fmt::Debug;
     use test_case::test_case;
+
+    static TEST_BYTES: [u8; 4] = [0xde, 0xad, 0xbe, 0xef];
+
+    #[test]
+    fn happy_regular_nibbles() {
+        let iter = NibblesIterator::new(&TEST_BYTES);
+        let expected = [0xd, 0xe, 0xa, 0xd, 0xb, 0xe, 0xe, 0xf];
+
+        assert!(iter.eq(expected));
+    }
+
+    #[test]
+    fn size_hint() {
+        let mut iter = NibblesIterator::new(&TEST_BYTES);
+        assert_eq!((8, Some(8)), iter.size_hint());
+        let _ = iter.next();
+        assert_eq!((7, Some(7)), iter.size_hint());
+    }
+
+    #[test]
+    fn backwards() {
+        let iter = NibblesIterator::new(&TEST_BYTES).rev();
+        let expected = [0xf, 0xe, 0xe, 0xb, 0xd, 0xa, 0xe, 0xd];
+        assert!(iter.eq(expected));
+    }
+
+    #[test]
+    fn nth_back() {
+        let mut iter = NibblesIterator::new(&TEST_BYTES);
+        assert_eq!(iter.nth_back(0), Some(0xf));
+        assert_eq!(iter.nth_back(0), Some(0xe));
+        assert_eq!(iter.nth_back(1), Some(0xb));
+        assert_eq!(iter.nth_back(2), Some(0xe));
+        assert_eq!(iter.nth_back(0), Some(0xd));
+        assert_eq!(iter.nth_back(0), None);
+    }
+
+    #[test]
+    fn empty() {
+        let nib = NibblesIterator::new(&[]);
+        assert!(nib.is_empty());
+        let it = nib.into_iter();
+        assert!(it.is_empty());
+        assert_eq!(it.size_hint().0, 0);
+    }
+
+    #[test]
+    fn not_empty_because_of_data() {
+        let mut iter = NibblesIterator::new(&[1]);
+        assert!(!iter.is_empty());
+        assert!(!iter.is_empty());
+        assert_eq!(iter.size_hint(), (2, Some(2)));
+        assert_eq!(iter.next(), Some(0));
+        assert!(!iter.is_empty());
+        assert_eq!(iter.size_hint(), (1, Some(1)));
+        assert_eq!(iter.next(), Some(1));
+        assert!(iter.is_empty());
+        assert_eq!(iter.size_hint(), (0, Some(0)));
+    }
 
     #[test_case([0, 0, 2, 3], [2, 3])]
     #[test_case([1, 2, 3, 4], [2, 3, 4])]
